Finished start/death screen
The game will now start with a menu screen telling the player to press space to play, then if they do and die it will end on the same screen instead showing the player score instead of the press space message. If the player presses space, the game will start again.